Created BeginTransactionTask. When the ProductListingActivity (products screen) is opened, it creates a transaction now on the server. getTransactions() is created added in TransactionService. Added ConfirmTransactionTask to handle updating the Transaction in the server after confirming the order in the client

+++ app/src/main/java/edu/uark/uarkregisterapp/models/api/Transaction.java
@@ -18,77 +18,63 @@ import edu.uark.uarkregisterapp.models.api.interfaces.LoadFromJsonInterface;
 
 public class Transaction implements ConvertToJsonInterface, LoadFromJsonInterface<Transaction> {
     private UUID recordId;
-
     public UUID getId() {
         return recordId;
     }
-
     public Transaction setId(UUID recordId) {
         this.recordId = recordId;
         return this;
     }
 
     private String cashierId; // todo check if it should be  a string to reference employeeId or a UUID
-
     public String getCashierId() {
         return cashierId;
     }
-
     public Transaction setCashierId(String cashierId) {
         this.cashierId = cashierId;
         return this;
     }
 
     private double totalAmount;
-
     public double getTotalAmount() {
         return totalAmount;
     }
-
-    public Transaction setTotalAmount(long totalAmount) {
+    public Transaction setTotalAmount(double totalAmount) {
         this.totalAmount = totalAmount;
         return this;
     }
 
     private String transactionType;
-
     public String getTransactionType() {
         return transactionType;
     }
-
     public Transaction setTransactionType(String transactionType) {
         this.transactionType = transactionType;
         return this;
     }
 
     private UUID referenceId; //change data type as needed
-
     public UUID getReferenceId() {
         return referenceId;
     }
-
     public Transaction setReferenceId(UUID referenceId) {
         this.referenceId = referenceId;
         return this;
     }
 
     private Date createdOn;
-
     public Date getCreatedOn() {
         return createdOn;
     }
-
     public Transaction setCreatedOn(Date createdOn) {
         this.createdOn = createdOn;
         return this;
     }
 
     private int totalItemSold; //total items sold may not be needed
-
     public int getTotalItemSold() {
         return totalItemSold;
     }
-
     public Transaction setTotalItemSold(int totalItemSold) {
         this.totalItemSold = totalItemSold;
         return this;
